% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_collex_fye.R
\name{collex_fye}
\alias{collex_fye}
\title{Iterative \emph{Fisher-Yates Exact} test for collexeme/collocational analysis}
\usage{
collex_fye(a, n_corpus = "total size of the corpus",
  n_coll = "total frequency of collexemes/collocates in the corpus as a whole",
  n_cxn = "total frequency of the construction/node word in the corpus as a whole",
  two_sided = FALSE, collstr_res = TRUE, float = 3,
  mpfr_precision = 500)
}
\arguments{
\item{a}{cell \code{a} in a 2-by-2 crosstabulation matrix (viz. representing the co-occurrence tokens of the levels of the variables. For instance, word-word co-occurrences, or word-construction co-occurrences).}

\item{n_corpus}{the total size (in word tokens) of the corpus.}

\item{n_coll}{the total frequency of the collexemes/collocates of the target construction/node word in the corpus.}

\item{n_cxn}{the total frequency of occurrence of the target construction/node word in the corpus.}

\item{two_sided}{logical; whether to perform one-sided test (\code{FALSE} -- Default) or two-sided (\code{TRUE}).}

\item{collstr_res}{logical; whether output the FYE \emph{p}-value as the Collostruction Strength value (\code{TRUE} -- the default) or just report the \emph{p}-value (\code{FALSE}).}

\item{float}{the floating digits of the Collostruction Strength. The default is \code{3}.}

\item{mpfr_precision}{a number indicating the maximal precision to be used in \bold{bits}. This is passed to the \code{precBits} argument of \code{\link[Rmpfr]{mpfr}}.}
}
\value{
Numeric vector of the same length as \code{a} interpreted as the Collostruction Strength of the construction/node word with the collexemes/collocates.
    Collostruction Strength is (i) the negative logarithm to the base of ten of the Fisher-Yates Exact test \emph{p}-value when \code{a} > \code{a_exp}, and (ii) the positive logarithm when \code{a} <= \code{a_exp}.
}
\description{
This is a vectorised wrapper for the \code{\link[stats]{dhyper}} function in the \code{stats} package.
    The implementation of the code is adapted from Gries (2012).
    \code{collex_fye} also provides a logical argument (i.e., \code{two_sided}) whose value is passed to the \code{alternative} argument of the embedded \code{\link[stats]{fisher.test}} if \code{two_sided} is \code{TRUE}.
}
