% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_collex_prepare.R
\name{collex_prepare}
\alias{collex_prepare}
\title{Prepare data for collexeme/collocates analysis}
\usage{
collex_prepare(list_output = "list output of 'colloc_default()'",
  span = NULL)
}
\arguments{
\item{list_output}{The list output of \code{colloc_default}.}

\item{span}{character; the context-window span user wants to focus on for the collexeme/collocates analysis.
For instance \code{"r1"}, \code{"l1"}, or a set of span windows as \code{c("r1", "l1")}.
The default is \code{NULL}, indicating that all span defined from running \code{colloc_default} will be considered.}
}
\value{
A tbl_df (tibble data frame)
}
\description{
This function is designed to handle the output of \code{colloc_default} to generate a tidy data required as input of \code{collex_fye} for performing collexeme/collocate analysis.
}
\examples{
\dontrun{
# do the collocate search
df <- colloc_default(corpus_path = orti_bali_path,
                     pattern = "^nuju$",
                     window = "b", # focusing on both left and right context window
                     span = 3) # retrieve 3 collocates to the left and right of the node

# prepare the collexeme analysis input tibble
# and select to focus on R1 and R2 collocates.
collex_tb <- collex_prepare(df, span = c("r1", "r2"))
}

}
