% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_colloc_default.R
\name{colloc_default}
\alias{colloc_default}
\title{Collocates retrieval for raw corpus text}
\usage{
colloc_default(corpus_path = NULL, corpus_list = NULL, pattern = NULL,
  window = "b", span = 3, word_split_regex = "([^a-zA-Z-]+|--)",
  case_insensitive = TRUE, to_lower_colloc = TRUE,
  tokenise_corpus_to_sentence = TRUE)
}
\arguments{
\item{corpus_path}{character strings of (full) filepath for the corpus text files in \code{.txt} plain-text format.}

\item{corpus_list}{a named list object containing elements constituting a corpus text.
The name of each element should correspond to the corpus file.
There can be more than one element (hence more than one corpus text) within this list object.}

\item{pattern}{regular expressions/exact patterns for the target pattern.}

\item{window}{window-span direction of the collocates: \code{"r"} ('\bold{right} of the node'), \code{"l"} ('\bold{left} of the node'), or the DEFAULT is \code{"b"} ('both \bold{left} and \bold{right} context-window').}

\item{span}{integer vector indicating the span of the collocate scope.}

\item{word_split_regex}{user-defined regular expressions to tokenise the corpus.
The default is to split at non alphabetic characters but retain hypen "-" as to maintain reduplication, for instance.
The regex for this default setting is \code{"([^éa-zA-Z-]+|--)"}.}

\item{case_insensitive}{whether the search pattern ignores case (TRUE -- the default) or not (FALSE).}

\item{to_lower_colloc}{whether to lowercase the retrieved collocates (TRUE -- default) or not (FALSE).}

\item{tokenise_corpus_to_sentence}{whether to tokenise the input corpus by sentence so that the script can handle the collocates for not crossing sentence boundary.
The default is \code{TRUE} and it uses \code{\link[stringi]{stri_split_boundaries}} to tokenise into sentence before further tokenising into word-tokens with \code{\link[stringr]{str_split}}.}
}
\value{
A list of three elements:
\itemize{
    \item A tibble of all words in the corpus including the sentence number;
    \item A tibble of all retrieved collocates, including their span position and sentence number;
    \item Regular expression object of the search pattern.
}
}
\description{
Collocates retrieval for raw corpus text
}
\examples{
\dontrun{
# do the collocate search
df <- colloc_default(corpus_path = orti_bali_path,
                     pattern = "^nuju$",
                     window = "b", # focusing on both left and right context window
                     span = 3) # retrieve 3 collocates to the left and right of the node
}
}
