% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_colloc_1.1.R
\name{colloc_leipzig_vector}
\alias{colloc_leipzig_vector}
\title{Vectorised collocate extraction}
\usage{
colloc_leipzig_vector(corpus.vect.path = NULL, corpus.vect = NULL,
  pattern = NULL, case.insensitive = TRUE, window = "r", span)
}
\description{
Vectorised collocate extraction
}
