% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_wordlist.R
\name{create_wordlist}
\alias{create_wordlist}
\title{Generate word-/regex-specific frequency list from Leipzig Corpora}
\usage{
create_wordlist(pattern = NULL, corpus_file_names = "corpus-file path",
  case_insensitive = TRUE)
}
\arguments{
\item{pattern}{the regular expressions/exact patterns for the target pattern/word}

\item{corpus_file_names}{gives the (i) file names of the corpus if they are in the working directory, or (ii) the complete file path to each of the Leipzig}

\item{case_insensitive}{whether case differences should be ignored (\code{TRUE} -- the default) or not (\code{FALSE})}
}
\value{
a tibble with three columns (i) \code{match}, (ii) \code{corpus_id}, and (iii) \code{n}, which is the count/token
}
\description{
The function generates a tibble of word-count for particular words/regexes for each supplied Leipzig corpus file
}
\examples{

# prepare the input
regex <- "\\\\bmemberi(kan)?\\\\b"
corpus.path <- corpus_files_path[1:2]

# generate the frequency count
create_wordlist(pattern = regex, corpus_file_names = corpus.path, case_insensitive = TRUE)

}
