% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_corpus_bali}
\alias{demo_corpus_bali}
\title{Balinese newspapers texts}
\format{An object of class \code{list} of length 10.}
\usage{
demo_corpus_bali
}
\description{
A random ten corpus files of Balinese newspapers texts. These corpus files are raw plain texts without any annotation and which have not been split by sentences.
}
\keyword{datasets}
