% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_corpus_id}
\alias{demo_corpus_id}
\title{Indonesian short stories}
\format{An object of class \code{list} of length 10.}
\usage{
demo_corpus_id
}
\description{
A random ten corpus files of Indonesian short stories retrieved from online blog. These corpus files are raw plain texts without any annotation and which have not been split by sentences.
}
\keyword{datasets}
