% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flist_mini}
\alias{flist_mini}
\title{Unit testing data 1}
\format{
An object of class \code{list} of length 2.
}
\usage{
flist_mini
}
\description{
A list containing two mini data frame as output for \code{freqlist_leipzig_all()}. This list is used for unit testing data.
}
\keyword{datasets}
