% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_wordlist.R
\name{freqlist_create_all}
\alias{freqlist_create_all}
\title{Frequency list of all words in a corpus}
\usage{
freqlist_create_all(split_regex = "([^a-zA-Z0-9-]+|--)",
  corpus_file_names = "one corpus file per operation",
  case_insensitive = TRUE)
}
\arguments{
\item{split_regex}{user-defined regular expressions to tokenise the corpus.}

\item{corpus_file_names}{full filepath to ONE Leipzig Corpora.}

\item{case_insensitive}{Logical; ignoring (\code{TRUE}) or maintaining (\code{FALSE}) the case when splitting the corpus into word token.}
}
\value{
A tibble of frequency list in descending order of the frequency.
}
\description{
Frequency list of all words in a corpus
}
\examples{
\dontrun{
wlist_all <- freqlist_create_all(corpus_file_names = corpus_files_path[1])
}
}
