% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_freqlist_leipzig_citation.R
\name{freqlist_leipzig_citation}
\alias{freqlist_leipzig_citation}
\title{Check full sentence-citation for a pattern}
\usage{
freqlist_leipzig_citation(pattern = NULL,
  corpus_file_names = corpus_files_path, case_insensitive = TRUE)
}
\arguments{
\item{pattern}{regular expressions/strings of the target words/patterns.}

\item{corpus_file_names}{complete file path of the Leipzig corpora to search for full citation of the words/patterns of interest.}

\item{case_insensitive}{whether to search case-insensitive pattern (\code{TRUE} -- the default) or not (\code{FALSE}).}
}
\value{
character vectors containing the info for 'corpus names', 'sentence number', and 'sentence match'.
}
\description{
A function to retrieve a full-sentence citation for a match/pattern. This function can be used to check the full sentence in which a word/pattern occur in a corpus.
    This can be a useful follow-up upon inspecting the wordlist results derived via \code{\link{freqlist_leipzig_each}} function.
}
