% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_freqlist_leipzig_each.R
\name{freqlist_leipzig_each}
\alias{freqlist_leipzig_each}
\title{Generate word-/regex-specific frequency list from the Leipzig Corpora}
\usage{
freqlist_leipzig_each(pattern = NULL, corpus_file_names = "corpus filepath",
  case_insensitive = TRUE)
}
\arguments{
\item{pattern}{the regular expressions/exact patterns for the target pattern/word whose frequency in a (set of) Leipzig Corpus file(s) you want to generate.}

\item{corpus_file_names}{gives the (i) file names of the corpus if they are in the working directory, or (ii) the complete file path to each of the Leipzig.}

\item{case_insensitive}{logical; whether case differences should be ignored (\code{TRUE} -- the default) or not (\code{FALSE}).}
}
\value{
a tibble with three columns (i) \code{match}, (ii) \code{corpus_id}, and (iii) \code{n}, which is the count/token.
}
\description{
The function generates a tibble of token-count for a particular word(s)/regex(es) for each supplied Leipzig corpus file.
}
\examples{
\dontrun{
# prepare the input
regex <- "\\\\bmemberi(kan)?\\\\b"
corpus.path <- corpus_files_path[1:2]

# generate the frequency count
freqlist_leipzig_each(pattern = regex,
                corpus_file_names = corpus.path,
                case_insensitive = TRUE)
}
}
