% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_wordlist.R
\name{freqlist_summarise}
\alias{freqlist_summarise}
\title{Summarise wordlist count across corpus}
\usage{
freqlist_summarise(df, group_var = "variable to group by",
  descending = TRUE)
}
\arguments{
\item{df}{a tibble data frame containing wordlist derived via the \code{\link{freqlist_create}}.}

\item{group_var}{variable to group by.}

\item{descending}{whether the list is ordered increasingly and vice versa according to the token frequency.}
}
\value{
a tibble
}
\description{
function to summarise the wordlist by removing the \code{corpus} variable derived via \code{\link{freqlist_create}}.
}
\examples{
\dontrun{
regex <- "\\\\bmemberi(kan)?\\\\b"
corpus.path <- corpus_files_path[1:2]

# Generate the freqlist of the pattern
wlist <- freqlist_create(pattern = regex,
                         corpus_file_names = corpus.path,
                         case_insensitive = TRUE)

wlist
A tibble: 4 x 3
match      corpus_id              n
<chr>      <chr>              <int>
 1 memberi    ind_mixed_2012_1M   6394
 2 memberikan ind_mixed_2012_1M  11710
 3 memberi    ind_news_2008_300K  2214
 4 memberikan ind_news_2008_300K  5213

# Summarise the match
freqlist_summarise(wlist, match, T)
}
}
