% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_word_ngram.R
\name{joint_freq}
\alias{joint_freq}
\title{Combine the ngram output with the frequency list of the elements of the ngram}
\usage{
joint_freq(df.x, df.y, ...)
}
\arguments{
\item{df.x}{ngram table output by \code{ngram_word_leipzig()}.}

\item{df.y}{word list table output by \code{ngram_word_leipzig()}.}

\item{...}{capture the word elements of the ngram (e.g., \code{w1, w2, w3, etc.}).}
}
\value{
a joint tibble for the ngram and frequency list for each component word of the ngram
}
\description{
Left join the frequency list for each element making up the ngram
}
\examples{

# load the corplingr package
devtools::load_all("~/Documents/_my_r_packages/corplingr")

# load the .RData for the cleaned sentence corpora
load(corpus_cleaned_path[1])

# if it is a sentence-per-line input-vector, mark the element with "SENT"
sentence_cleaned <- paste("SENT", sentence_cleaned, "SENT", sep = " ")

# generate the ngram
ngramout <- ngram_word_leipzig(sentence_cleaned, sample_n = 10000)
ngram <- ngramout$ngram
wlist <- ngramout$wlist

# joint the frequency list
df <- joint_freq(ngram, wlist, w1, w2)
df

}
