% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_wordlist.R
\name{leipzig_corpus_size}
\alias{leipzig_corpus_size}
\title{Generate Leipzig corpus-size}
\usage{
leipzig_corpus_size(word_regex = "\\\\b(?i)([-a-z0-9]+)\\\\b",
  corpus_file_names = corpus_files_path)
}
\arguments{
\item{word_regex}{regular expressions defining what "a word" is}

\item{corpus_file_names}{file path to the directory folder in which the Leipzig corpus files are stored}
}
\value{
tibble containing \code{corpus_id}, \code{size}, and \code{size_print} (for text-printing)
}
\description{
function to get a total word-token count of a given leipzig corpus file
}
