% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_leipzig_1}
\alias{mini_leipzig_1}
\title{Unit testing data 3}
\format{
An object of class \code{character} of length 30.
}
\usage{
mini_leipzig_1
}
\description{
A plain text containing thirty random sentences from the Indonesian Leipzig corpora. This file is stored in the testthat directory and used for unit testing data.
}
\keyword{datasets}
