% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_ngram_word_jointfreq.R
\name{ngram_word_jointfreq}
\alias{ngram_word_jointfreq}
\title{Combine the ngram output with the frequency list of the elements of the ngram}
\usage{
ngram_word_jointfreq(df_ngram, df_wlist, ...)
}
\arguments{
\item{df_ngram}{ngram table output by \code{\link{ngram_word_leipzig}}.}

\item{df_wlist}{wordlist table output by \code{\link{ngram_word_leipzig}}.}

\item{...}{capture the word elements of the ngram (e.g., \code{w1, w2, w3, etc.}).}
}
\value{
a joint tibble for the ngram and frequency list for each component word of the ngram
}
\description{
Left join the frequency list for each element making up the ngram
}
\examples{
\dontrun{
# load the corplingr package
devtools::load_all("~/Documents/_my_r_packages/corplingr")

# load the .RData for the cleaned sentence corpora
# 'cleaned' -> all tokens have been separated by whitespace
load(leipzig_cleaned_path[1])

# if it is a sentence-per-line input-vector, mark the element with "SENT"
sentence_cleaned <- paste("SENT", sentence_cleaned, "SENT", sep = " ")

# generate the ngram
ngram_out <- ngram_word_leipzig(sentence_cleaned, sample_n = 10000)
ngram <- ngram_out$ngram
wlist <- ngram_out$wlist

# joint the frequency list
df <- ngram_word_jointfreq(ngram, wlist, w1, w2)
df
}
}
