% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obali_colloc_output_test}
\alias{obali_colloc_output_test}
\title{Unit testing data 4}
\format{
An object of class \code{list} of length 3.
}
\usage{
obali_colloc_output_test
}
\description{
A list containing outputs from \code{colloc_default()}. This list is used for unit testing data.
}
\keyword{datasets}
