% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_span_set.R
\name{span_set}
\alias{span_set}
\title{Prepare window-span vector for collocates}
\usage{
span_set(window = "character", span = "integer")
}
\arguments{
\item{window}{window-span direction of the collocates: \code{"r"} ('\bold{right} of the node'), \code{"l"} ('\bold{left} of the node'), or the DEFAULT is \code{"b"} ('both \bold{left} and \bold{right} context-window').}

\item{span}{integer vector indicating the span of the collocate scope.}
}
\value{
Named integer vectors of window-span. The name attributes refer to the window position of the vector (e.g. "l2", "l1", "node", "r1", "r2").
}
\description{
Utility internal-function for constructing window-span vector in the collocate function.
}
\examples{
span <- span_set(window = "b", span = 3)
}
