% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_sent_tokeniser.R
\name{tokenise_sentence}
\alias{tokenise_sentence}
\title{Split a corpus by sentence-boundary}
\usage{
tokenise_sentence(strings = NULL, to_lower = TRUE,
  window_span = NULL)
}
\arguments{
\item{strings}{character vector of a corpus text.}

\item{to_lower}{logical; turn the corpus into lowercase when \code{TRUE} (the default).}

\item{window_span}{integer; it is supplied from the value of the \code{span} argument in the higher-level collocational function call.
It will determine the number of times the \code{"ZSENTENCEZ"} marker will be appended at the beginning and at the end of each sentence.}
}
\value{
A character vector with as many sentences as there are in the input corpus as identified by \code{\link[stringi]{stri_split_boundaries}}.
}
\description{
The embedded function in the collocational framework to split input corpus into vector of sentences using \code{\link[stringi]{stri_split_boundaries}} from \code{stringi} package.
    Each sentence line will be appended, at the beginning and at the end, with \code{"ZSENTENCEZ"} marker as many as the number of collocational window-span is required.
    This marker will help identify if collocates of a word cross the boundary of the sentence in which the word occurs.
    The function automatically detects and removes if \code{"ZSENTENCEZ"} is part of the identified collocate.
}
\examples{
txt <- c("It is one sentence. It is another sentence! There are TWO sentences.",
         "The second sentence and the second element of 'txt' corpus.",
         "Can we add another one (sentence) here as the third element?",
         "Of course!")
sent <- tokenise_sentence(strings = txt,
                          to_lower = TRUE,
                          window_span = 3)
sent
}
