% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_wordlist.R
\name{wordlist_leipzig_read}
\alias{wordlist_leipzig_read}
\title{Read in the wordlist files of the Leipzig Corpora}
\usage{
wordlist_leipzig_read(wlist_path = wordlist_corpus_path,
  file_pattern = NULL, all = FALSE, lower_case = TRUE)
}
\arguments{
\item{wlist_path}{the file paths of the wordlist. Use \code{wordlist_files_path}}

\item{file_pattern}{the regular expressions/ID labels for the corpus files}

\item{all}{whether to read all available wordlist files from all corpus (\code{TRUE}) or not (\code{FALSE} -- the default)}

\item{lower_case}{whether the words are lowercased (\code{TRUE} -- the default) or not (\code{FALSE})}
}
\value{
a tibble with three columns (i) \code{corpus_id} (i.e., the corpus file), (ii) \code{word} (i.e., word forms), and (iii) \code{n} (i.e., token frequency)
}
\description{
The function generates a tibble of wordlist read from the Leipzig Corpora Word List files
}
