% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_wordlist.R
\name{wordlist_summarise}
\alias{wordlist_summarise}
\title{Summarise wordlist count across corpus}
\usage{
wordlist_summarise(df, group_var = "variable to group_by",
  descending = TRUE)
}
\arguments{
\item{df}{a tibble data frame containing wordlist derived via the \code{\link{create_wordlist()}}.}

\item{descending}{whether the list is ordered increasingly and vice versa according to the token frequency}
}
\value{
a tibble
}
\description{
function to summarise the wordlist by removing the \code{corpus} variable derived via \code{\link{create_wordlist()}}.
}
