% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_colloc_leipzig.R
\name{colloc_leipzig}
\alias{colloc_leipzig}
\title{Generate tidyverse-style window-span collocates for the Leipzig Corpora}
\usage{
colloc_leipzig(corpus = "path to leipzig corpus", pattern = "regex",
  window = "b", span = 2, case_insensitive = TRUE,
  tolower_colloc = TRUE, save_results = FALSE,
  coll_output_name = "colloc_tidy_colloc_out.txt",
  sent_output_name = "colloc_tidy_sent_out.txt")
}
\arguments{
\item{corpus}{(i) file names of the corpus if they are in the working directory, or (ii) the complete file path to each of the Leipzig corpus files}

\item{pattern}{regular expressions/exact patterns for the target pattern}

\item{window}{window-span direction of the collocates: \code{"r"} ('right of the node'), \code{"l"} ('left of the node'), or the DEFAULT is \code{"b"} ('both window')}

\item{span}{numeric vector indicating the span of the collocate scope}

\item{case_insensitive}{whether the search ignores case (TRUE -- the default) or not (FALSE)}

\item{tolower_colloc}{whether to lower case the retrieved collocates and the nodes (TRUE -- default) or not (FALSE)}

\item{save_results}{whether to output the collocates into a tab-separated plain text (TRUE) or not (FALSE -- default)}

\item{coll_output_name}{name of the file for the collocate tables}

\item{sent_output_name}{name of the file for the full sentence match containing the collocates}
}
\value{
a list of two tibbles: (i) for collocates with sentence number of the match, window span information, and the corpus files, and (ii) full-sentences per match with sent number and corpus file
}
\description{
The function produces tibble-output collocates for Leipzig Corpora files.
}
\examples{
\dontrun{
colloc <- colloc_leipzig(corpus = corpus_files_path[2:3],
                              pattern = "\\\\bterelakkan\\\\b",
                              window = "b",
                              span = 3,
                              save_results = FALSE,
                              tolower_colloc = TRUE)
# Inspect outputs

colloc$collocates

colloc$sentence_matches
}
}
