% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_corpus_size_leipzig.R
\name{corpus_size_leipzig}
\alias{corpus_size_leipzig}
\title{Generate Leipzig corpus-size}
\usage{

  corpus_size_leipzig(leipzig_path = "(full) filepath to Leipzig corpus files",
  word_regex = "\\\\b(?i)([-a-zA-Z0-9]+)\\\\b")
}
\arguments{
\item{leipzig_path}{file path to the directory folder in which the Leipzig corpus files are stored}

\item{word_regex}{regular expressions defining what "a word" is}
}
\value{
tibble containing \code{corpus_id}, \code{size}, and \code{size_print} (for text-printing)
}
\description{
function to get a total word-token count of a given leipzig corpus file.
    It is built on top of \code{\link[stringr]{str_count}}.
}
