% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplingr_fye.R
\name{fye2}
\alias{fye2}
\title{Fisher Yates Exact test}
\usage{
fye2(a, b, c, d, a.exp, two.sided = FALSE, res = c("p", "or"), float = 3)
}
\arguments{
\item{a}{cell \code{a} in a 2-by-2 crosstabulation matrix (viz. representing the co-occurrence tokens of the levels of the variables. For instance, word-word co-occurrences, or word-construction co-occurrences).}

\item{b}{cell \code{b} in the matrix (it represents alternative items for the co-occurrence of the first row and the second column. For instance, alternative construction, or near-synonyms).}

\item{c}{cell \code{c} in the matrix (i.e., intersection of the second row and the first column (\code{[2, 1]}))}

\item{d}{cell \code{d} in the matrix (i.e., intersection of the second row and the second column (\code{[2, 2]}))}

\item{a.exp}{expected frequency in the reference cell \code{a} of the 2-by-2 crosstabulation matrix.}

\item{two.sided}{whether to perform one-sided test (\code{FALSE} -- Default) or two-sided (\code{TRUE}).}

\item{res}{choice of outputs: (i) the \code{p.value} ("p") or (ii) \code{odds.ratio} ("or").}

\item{float}{the floating digits (\code{3} -- Default).}
}
\value{
vector of the same length as \code{a}.
}
\description{
This is a vectorised wrapper for \code{fisher.test()} in the \code{stats} package. The implementation of the code is slightly adapted from Levshina's (2015) \code{pv.Fisher.collstr()}. \code{fye2()} adds the argument for values in the \code{alternative} argument in the \code{fisher.test()}.
}
